requirements.txt
requests
redis
bs4
jinja2
pandas
dnspython
